requirements.txt
streamlit>=1.36.0
pandas>=2.0.0
numpy>=1.24.0
xgboost>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
plotly>=5.18.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pytz>=2023.3
python-dotenv>=1.0.0
fpdf2>=2.7.0
schedule>=1.2.0
torch>=2.0.0
cloudscraper>=1.2.71
